🧹 chore: strip menu-bar slop — drop pid-file contents, debug flag, no-value tests
Trims the menu-bar PR down to just what's needed to show the icon:

- Remove writePID/readPID and rename menu.pid → menu.lock. The flock is the
  source of truth for liveness; the PID contents were write-only, never read
  at runtime (only by tests).
- Drop the unused --debug pass-through in Spawn — menu.go never reads it.
- Drop configDir field on ServeCommander, plumb directly from RunE.
- Delete the empty-function panic test, the PNG-decode test (compile-time
  embed catches this), and the Spawn-doesn't-panic test.
- Drop the unrelated /.worktrees/ .gitignore addition.
