+++ .gitignore
@@ -20,6 +20,3 @@
 
 # Claude Code
 .claude/
-
-# Git worktrees
-/.worktrees/
