+++ cmd/tapes/menu/menu.go
@@ -7,7 +7,6 @@ package menucmder
 import (
 	_ "embed"
 	"fmt"
-	"os"
 
 	"github.com/getlantern/systray"
 	"github.com/spf13/cobra"
@@ -24,10 +23,6 @@ func NewMenuCmd() *cobra.Command {
 		RunE: func(cmd *cobra.Command, _ []string) error {
 			configDir, _ := cmd.Flags().GetString("config-dir")
 
-			// Acquire the menu lock. If held by another live instance we
-			// exit silently — the existing menu owns the menu bar slot.
-			// The kernel auto-releases the flock when this process exits,
-			// so a SIGKILL cannot strand the lock.
 			lock, err := acquireMenuLock(configDir)
 			if err != nil {
 				return fmt.Errorf("acquiring menu lock: %w", err)
@@ -36,12 +31,6 @@ func NewMenuCmd() *cobra.Command {
 				return nil
 			}
 			defer func() { _ = lock.Close() }()
-
-			if err := writePID(lock, os.Getpid()); err != nil {
-				// Informational only — proceed even if we cannot persist
-				// the pid for human inspection.
-				_ = err
-			}
 
 			systray.Run(onReady, onExit)
 			return nil
