+++ cmd/tapes/menu/spawn_other.go
@@ -5,4 +5,4 @@ package menucmder
 import "log/slog"
 
 // Spawn is a no-op on non-darwin platforms.
-func Spawn(_ string, _ bool, _ *slog.Logger) {}
+func Spawn(_ string, _ *slog.Logger) {}
